Remove boot-service.qti.xml during boot into recovery mode
* Fix the recovery bootloop issue in new A15 QPR2 related ROMs

* Some QPR2 ROMS made changes to the used boot control HAL.
The corresponding manifest definition got placed in /system/etc/vintf/manifest/boot-service.qti.xml, which is located in vendor_boot and which TWRP doesn't control.
That manifest has a) an incompatible version 9.0 and TWRP (Android 12.1) uses 4.0 and b) it is flagged as a device manifest.
Flagging it as a device manifest is correct, but it should have been placed in /vendor instead of /system, according to official Android docs.
The fix for now is to remove that manifest file (in memory) during boot into recovery mode, because TWRP provides its own version.

Signed-off-by: HDzungx <hdzungx@gmail.com>